Move PlanetTagInputDialog to separate file
