+++ src/app/shared/forms/planet-forms.module.ts
@@ -8,9 +8,10 @@ import { FormErrorMessagesComponent } from './form-error-messages.component';
 import { PlanetRatingComponent } from './planet-rating.component';
 import { PlanetRatingStarsComponent } from './planet-rating-stars.component';
 import { PlanetStackedBarComponent } from './planet-stacked-bar.component';
-import { PlanetTagInputComponent, PlanetTagInputDialogComponent } from './planet-tag-input.component';
+import { PlanetTagInputComponent } from './planet-tag-input.component';
 import { PlanetStepListComponent, PlanetStepListFormDirective, PlanetStepListItemComponent } from './planet-step-list.component';
 import { PlanetMarkdownTextboxComponent } from './planet-markdown-textbox.component';
+import { PlanetTagInputDialogComponent } from './planet-tag-input-dialog.component';
 
 @NgModule({
   imports: [
